REFACTOR : JoinInfoRepository 경로 변경
